Refactor to fix celery auto discovery
