+++ gazebo/supervisord.conf
@@ -12,7 +12,7 @@ stdout_logfile_backups=0
 stderr_logfile_backups=0
 
 [program:celery-app]
-command=celery -A simulator_api.celery_instance.celery.celery_instance worker --loglevel=info
+command=celery -A simulator_api.celery_tasks.tasks.celery_instance worker --loglevel=info
 user=movai
 stderr_logfile=/dev/fd/1
 stdout_logfile=/dev/fd/1
